requirements.txt
# Core dependencies
flask==2.0.1
numpy==1.21.0
Pillow==8.3.1
gunicorn==20.1.0

# Optional dependencies (will be used if available)
scikit-learn>=1.0.2; platform_machine != "armv7l"

# Build tools
wheel==0.37.1
